Move artifacts modal files to JobArtifacts subfolder
Move ViewJobArtifactsTabbedModal, ViewArtifactsModal, ViewJobDatasetsModal,
ViewJobModelsModal, and ArtifactPreviewPane into Tasks/JobArtifacts/.
Update all import paths accordingly.
