+++ src/renderer/components/Experiment/Tasks/JobArtifacts/ArtifactPreviewPane.tsx
