+++ src/renderer/components/Experiment/Tasks/JobArtifacts/ViewArtifactsModal.tsx
