+++ src/renderer/components/Experiment/Tasks/Tasks.tsx
@@ -27,16 +27,16 @@ import EditInteractiveTaskModal from './EditInteractiveTaskModal';
 import DeleteTaskConfirmModal from './DeleteTaskConfirmModal';
 import QueueTaskModal from './QueueTaskModal';
 import ViewOutputModalStreaming from './ViewOutputModalStreaming';
-import ViewArtifactsModal from './ViewArtifactsModal';
+import ViewArtifactsModal from './JobArtifacts/ViewArtifactsModal';
 import ViewProfilingModal from './ViewProfilingModal';
 import ViewCheckpointsModal from './ViewCheckpointsModal';
 import ViewEvalResultsModal from './ViewEvalResultsModal';
 import CompareEvalResultsModal from './CompareEvalResultsModal';
 import PreviewDatasetModal from '../../Data/PreviewDatasetModal';
 import ViewSweepResultsModal from './ViewSweepResultsModal';
-import ViewJobDatasetsModal from './ViewJobDatasetsModal';
-import ViewJobModelsModal from './ViewJobModelsModal';
-import ViewJobArtifactsTabbedModal from './ViewJobArtifactsTabbedModal';
+import ViewJobDatasetsModal from './JobArtifacts/ViewJobDatasetsModal';
+import ViewJobModelsModal from './JobArtifacts/ViewJobModelsModal';
+import ViewJobArtifactsTabbedModal from './JobArtifacts/ViewJobArtifactsTabbedModal';
 import FileBrowserModal from './FileBrowserModal';
 import SafeJSONParse from '../../Shared/SafeJSONParse';
 import NewTaskModal2 from './NewTaskModal/NewTaskModal2';
